rename math.h to extended_math.h
