+++ internal/c/libqb/include/extended_math.h
